backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
python-pptx==0.6.23
pydantic==2.5.3
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
aiofiles==23.2.1
openai==1.12.0
jinja2==3.1.3
